Sleigh Construction Table Functionality part 3 (Hide menu for now)

+++ src/main/java/holiday_mod/world/item/SleighItem.java
@@ -68,7 +68,7 @@ public class SleighItem extends Item {
                 Vec3 vec31 = pPlayer.getEyePosition();
 
                 for(Entity entity : list) {
-                    AABB aabb = entity.getBoundingBox().inflate((double)entity.getPickRadius());
+                    AABB aabb = entity.getBoundingBox().inflate(entity.getPickRadius());
                     if (aabb.contains(vec31)) {
                         return InteractionResultHolder.pass(itemstack);
                     }
